requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.4
django-js-asset==1.2.2
gunicorn==20.0.4
numpy==1.18.2
psycopg2-binary==2.8.5
PyPrind==2.11.2
pytils==0.3
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
django-ckeditor==5.9.0
